requirements.txt
selenium
lxml
vk
googlesearch
colorama
pypdf2
python-Levenshtein
python-dateutil
bs4
networkx
matplotlib
